Refactor Transaction Service
—Medhat, Mohamed Zarea and Mohamed Sallam (Pair-programming)

+++ src/main/java/eg/boe/banqueofegypt/entity/Status.java
@@ -3,12 +3,12 @@ package eg.boe.banqueofegypt.entity;
 import java.util.Map;
 
 public enum Status {
-    PENDING(0), SUCCESS(200),
-    INSUFFICIENT_FUNDS(80090), INVALID_DATA(99993),
-    DEST_TIMEOUT(80091), SRC_TIMEOUT(80092),
-    FAILED(404);
+    PENDING(0L), SUCCESS(200L),
+    INSUFFICIENT_FUNDS(80090L), INVALID_DATA(99993L),
+    DEST_TIMEOUT(80091L), SRC_TIMEOUT(80092L),
+    FAILED(404L);
 
-    public final static Map<Integer, String> codeMessageMapper = Map.of(
+    public final static Map<Long, String> codeMessageMapper = Map.of(
             PENDING.code, "Pending",
             SUCCESS.code, "Success",
             INSUFFICIENT_FUNDS.code, "Insufficient funds",
@@ -17,9 +17,9 @@ public enum Status {
             SRC_TIMEOUT.code, "Source timeout"
     );
 
-    public final Integer code;
+    public final Long code;
 
-    Status(Integer code) {
+    Status(Long code) {
         this.code = code;
     }
 }
